# feat: Add E2E smoke test categorization for faster development feedback