--- a/package.json
+++ b/package.json
@@ -16,6 +16,8 @@
     "postinstall": "if [ \"$NODE_ENV\" != \"production\" ]; then playwright install chromium; fi",
     "test:unit": "if [ \"$NODE_ENV\" != \"production\" ]; then jest; else echo 'Skipping tests in production'; fi",
     "test:e2e": "if [ \"$NODE_ENV\" != \"production\" ]; then playwright test; else echo 'Skipping E2E tests in production'; fi",
+    "test:e2e:smoke": "if [ \"$NODE_ENV\" != \"production\" ]; then playwright test --grep @smoke; else echo 'Skipping E2E tests in production'; fi",
+    "test:e2e:critical": "if [ \"$NODE_ENV\" != \"production\" ]; then playwright test --grep @critical; else echo 'Skipping E2E tests in production'; fi",
     "test": "if [ \"$NODE_ENV\" != \"production\" ]; then npm run test:unit; else echo 'Skipping tests in production'; fi",
     "security": "eslint --config eslint.config.js src/ --rule 'security/*: error'",
     "security:fix": "eslint --config eslint.config.js src/ --fix",