## Summary

Add smoke test categorization system to E2E tests enabling faster local development feedback while maintaining comprehensive CI coverage.

## Changes Made

### Test Categorization
- **Added test tags**: `@smoke` and `@critical` to key E2E workflows
- **Tagged tests**:
  - `@critical @smoke`: Page load, note CRUD operations, note connections  
  - `@smoke`: Multi-select notes, export functionality

### New npm Scripts
```bash
npm run test:e2e:smoke      # ~5 core workflows (~13 seconds)
npm run test:e2e:critical   # ~3 essential flows (~2.5 seconds) 
npm run test:e2e            # All 21 tests (~2 minutes)
```

### Documentation Updates
- Updated `docs/testing.md` with test categorization strategy
- Added usage guidance for different development scenarios
- Updated `CLAUDE.md` with task management workflow instructions

### Strategy
- **CI/GitHub**: Continue running full test suite (browser install is the bottleneck)
- **Local Development**: Use smoke tests for rapid iteration and feedback
- **Flexibility**: Developers can choose appropriate test scope based on needs

## Test Results

✅ **Smoke tests**: 5 tests in ~13 seconds  
✅ **Critical tests**: 3 tests in ~2.5 seconds  
✅ **Full suite**: All 21 tests continue to work  

## Benefits

- **Faster feedback** during local development (13s vs 2min)
- **Maintained coverage** in CI with full test suite
- **Developer choice** between speed and comprehensiveness
- **Clear categorization** of test importance and scope

🤖 Generated with [Claude Code](https://claude.ai/code)

Co-Authored-By: Claude <noreply@anthropic.com>